[develop] 캘런더 버그 수정

+++ app/src/main/java/com/devidea/aicar/ui/main/components/history/HistoryComponent.kt
@@ -80,7 +80,7 @@ fun HistoryRoute(
             onSessionClick(id)
         },
         onDateSelected = { date -> viewModel.selectDate(date) },
-        onMonthChanged = { month -> viewModel.changeMonth(month.toInt()) },
+        onMonthChanged = { month -> viewModel.changeMonth(month) },
         onDeleteSessions = { sessionIds ->
             sessionIds.forEach { viewModel.deleteSession(it) }
         },
